requirements.txt
flask
flask-cors
flask-jwt-extended
psycopg2-binary
python-dotenv
RouterOS-api
sqlalchemy
sqlmodel
user-agents
uvicorn
